backend/requirements.txt
# ── Core ──────────────────────────────────────────────────
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.10,<3.0
pydantic-settings>=2.7,<3.0
python-multipart>=0.0.18

# ── Database ──────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.30,<1.0
psycopg2-binary>=2.9,<3.0
alembic>=1.14,<2.0

# ── Auth ──────────────────────────────────────────────────
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
bcrypt>=4.2,<5.0
twilio>=9.0,<10.0

# ── Cache / Queue ─────────────────────────────────────────
redis>=5.2,<6.0

# ── Utilities ─────────────────────────────────────────────
httpx>=0.28,<1.0
python-dotenv>=1.0,<2.0

# ── Testing ───────────────────────────────────────────────
pytest>=8.0,<9.0
pytest-asyncio>=0.25,<1.0
httpx  # test client
